+++ src/charm.py
@@ -23,7 +23,12 @@ from cosl.rules import AlertRules
 from ops import main
 from ops.model import BlockedStatus, MaintenanceStatus, Relation
 
-from grafana_agent import CONFIG_PATH, METRICS_RULES_SRC_PATH, GrafanaAgentCharm
+from grafana_agent import (
+    CONFIG_PATH,
+    METRICS_RULES_SRC_PATH,
+    GrafanaAgentCharm,
+    key_value_pair_string_to_dict,
+)
 from snap_management import SnapSpecError, install_ga_snap
 
 logger = logging.getLogger(__name__)
@@ -31,37 +36,6 @@ logger = logging.getLogger(__name__)
 _FsType = str
 _MountOption = str
 _MountOptions = List[_MountOption]
-
-
-def key_value_pair_string_to_dict(key_value_pair: str) -> dict:
-    """Transform a comma-separated key-value pairs into a dict."""
-    result = {}
-
-    for pair in key_value_pair.split(","):
-        pair = pair.strip()
-        if not pair:
-            continue
-
-        if ":" in pair:
-            sep = ":"
-        elif "=" in pair:
-            sep = "="
-        else:
-            logger.error("Invalid pair without separator ':' or '=': '%s'", pair)
-            continue
-
-        key, value = map(str.strip, pair.split(sep, 1))
-
-        if not key:
-            logger.error("Empty key in pair: '%s'", pair)
-            continue
-        if not value:
-            logger.error("Empty value in pair: '%s'", pair)
-            continue
-
-        result[key] = value
-
-    return result
 
 
 def inject_extra_labels_to_alert_rules(rules: dict, extra_alert_labels: dict) -> dict:
