+++ src/grafana_agent.py
@@ -57,6 +57,37 @@ DASHBOARDS_DEST_PATH = "grafana_dashboards"  # placeholder until we figure out t
 RulesMapping = namedtuple("RulesMapping", ["src", "dest"])
 
 
+def key_value_pair_string_to_dict(key_value_pair: str) -> dict:
+    """Transform a comma-separated key-value pairs into a dict."""
+    result = {}
+
+    for pair in key_value_pair.split(","):
+        pair = pair.strip()
+        if not pair:
+            continue
+
+        if ":" in pair:
+            sep = ":"
+        elif "=" in pair:
+            sep = "="
+        else:
+            logger.error("Invalid pair without separator ':' or '=': '%s'", pair)
+            continue
+
+        key, value = map(str.strip, pair.split(sep, 1))
+
+        if not key:
+            logger.error("Empty key in pair: '%s'", pair)
+            continue
+        if not value:
+            logger.error("Empty value in pair: '%s'", pair)
+            continue
+
+        result[key] = value
+
+    return result
+
+
 class GrafanaAgentReloadError(Exception):
     """Custom exception to indicate that grafana agent config couldn't be reloaded."""
 
@@ -153,11 +184,15 @@ class GrafanaAgentCharm(CharmBase):
                 shutil.copytree(rules.src, rules.dest, dirs_exist_ok=True)
 
         alert_rules_path = self.metrics_rules_paths.dest
+        extra_alert_labels = key_value_pair_string_to_dict(
+            cast(str, self.model.config.get("extra_alert_labels", ""))
+        )
         self._remote_write = PrometheusRemoteWriteConsumer(
             self,
             alert_rules_path=alert_rules_path,
             forward_alert_rules=self._forward_alert_rules,
             refresh_event=[self.on.config_changed],
+            extra_alert_labels=extra_alert_labels,
         )
 
         self._loki_consumer = LokiPushApiConsumer(
